Reset project + add Supabase configuration
