+++ kaapeh-house-app/.env.example
@@ -1,0 +1,5 @@
+﻿# Supabase Configuration
+# Copy this file to .env by running `copy .env.example .env.local` in Terminal
+# Then, replace with actual Supabase project values
+EXPO_PUBLIC_SUPABASE_URL=your_supabase_project_url_here
+EXPO_PUBLIC_SUPABASE_ANON_KEY=your_supabase_anon_key_here
