+++ kaapeh-house-app/package.json
@@ -11,10 +11,14 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^15.0.2",
+    "@react-native-async-storage/async-storage": "2.2.0",
     "@react-navigation/bottom-tabs": "^7.4.0",
     "@react-navigation/elements": "^2.6.3",
     "@react-navigation/native": "^7.1.8",
     "@react-navigation/native-stack": "^7.3.16",
+    "@rneui/base": "^4.0.0-rc.8",
+    "@rneui/themed": "^4.0.0-rc.8",
+    "@supabase/supabase-js": "^2.58.0",
     "expo": "^54.0.1",
     "expo-asset": "~12.0.3",
     "expo-blur": "~15.0.2",
@@ -27,6 +31,7 @@
     "expo-status-bar": "~3.0.3",
     "expo-symbols": "~1.0.2",
     "expo-web-browser": "~15.0.2",
+    "postgres": "^3.4.7",
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-native": "0.81.4",
@@ -34,6 +39,8 @@
     "react-native-reanimated": "^4.1.0",
     "react-native-safe-area-context": "~5.6.0",
     "react-native-screens": "~4.16.0",
+    "react-native-url-polyfill": "^3.0.0",
+    "react-native-vector-icons": "^10.3.0",
     "react-native-web": "^0.21.0",
     "react-native-worklets": "0.5.1"
   },
@@ -42,6 +49,7 @@
     "@types/react": "~19.1.0",
     "eslint": "^9.25.0",
     "eslint-config-expo": "~10.0.0",
+    "supabase": "^2.47.2",
     "typescript": "~5.9.2"
   },
   "private": true
