+++ kaapeh-house-app/src-example/components/HapticTab.tsx
