+++ kaapeh-house-app/src-example/components/ParallaxScrollView.tsx
