+++ kaapeh-house-app/src-example/components/ui/TabBarBackground.tsx
