+++ kaapeh-house-app/src-example/constants/Colors.ts
