+++ kaapeh-house-app/src-example/navigation/index.tsx
@@ -1,0 +1,75 @@
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { createStaticNavigation, StaticParamList } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import { Platform } from 'react-native';
+
+import { Explore } from './screens/Explore';
+import { Home } from './screens/Home';
+import { NotFound } from './screens/NotFound';
+
+import { HapticTab } from '@/components/HapticTab';
+import { IconSymbol } from '@/components/ui/IconSymbol';
+import TabBarBackground from '@/components/ui/TabBarBackground';
+
+import '@expo/metro-runtime';
+
+const HomeTabs = createBottomTabNavigator({
+  screens: {
+    Home: {
+      screen: Home,
+      options: {
+        headerShown: false,
+        tabBarIcon: ({ color }) => <IconSymbol size={28} name="house.fill" color={color} />,
+      },
+    },
+    Explore: {
+      screen: Explore,
+      options: {
+        headerShown: false,
+        tabBarIcon: ({ color }) => <IconSymbol size={28} name="paperplane.fill" color={color} />,
+      },
+    },
+  },
+  screenOptions: {
+    headerShown: false,
+    tabBarButton: HapticTab,
+    tabBarBackground: TabBarBackground,
+    tabBarStyle: Platform.select({
+      ios: {
+        // Use a transparent background on iOS to show the blur effect
+        possition: 'absolute',
+      },
+      default: {},
+    }),
+  },
+});
+
+const RootStack = createNativeStackNavigator({
+  screens: {
+    HomeTabs: {
+      screen: HomeTabs,
+      options: {
+        headerShown: false,
+      },
+    },
+    NotFound: {
+      screen: NotFound,
+      options: {
+        title: '404',
+      },
+      linking: {
+        path: '*',
+      },
+    },
+  },
+});
+
+export const Navigation = createStaticNavigation(RootStack);
+
+type RootStackParamList = StaticParamList<typeof RootStack>;
+
+declare global {
+  namespace ReactNavigation {
+    interface RootParamList extends RootStackParamList {}
+  }
+}
