+++ kaapeh-house-app/src-example/navigation/screens/Explore.tsx
