+++ kaapeh-house-app/src-example/scripts/reset-project.js
